RESPLANDISSANT0 (#1)
* Removed no-margin from contributer header

* Added parsing to name in profile

* Fix isPackageInstalled false positives

* Fixing bug #139 - Moving from message dropdown to full view

* Changed socket URI

* #164 tweaks

* Removed debug code in notifications.js

* Fixed spacing in inbox.js

* Removed unneeded code from notifications.js

* Fixed modal position on mobile devices

* Fixed modal position on mobile devices

* Live levels class

* Bug fix for incorrect level checking code

* Added block for users with score 0 and karma < 0

* Fixed level checks with actual answers

* Fixed modal position

* Updated API access

* Added ability to remove flags from admin panel. Flags are now marked as valid or invalid and not simply removed

* Added clickable links to admin sections

* Added stats to forum sidebar

* Fixed merge conflict interaction.scss

* Logout users when account is deleted

* Removed call to WeChal from level complete

* Set position absolute for modal in mobile view

* Added wechall back in to levels script

* Remove invalid co

* Fixed profile history links to forum posts to use GET variable instead of hash

* Update terms.php

Spelling Correction.
Paragraph 9.  Disclaimer of Warranties, the word "MERCHANTIBILITY", should be spelled MERCHANTABILITY.

* Fix a little spelling mistake.

* Update register.php

Error Correction:   Line 28:  All members must obied by"  contains an error, it should be "abide by"

* Fix a little spelling mistake.

* Add LDAP as the primary authentication model. MySQL kept as backup

* Fixed mathematical error in class.cache.php

* Fixed SSL error when connecting to paypal via curl

* Limit site news to only be visible on the homepage for 2 weeks

* Removing old elements from home articles

* Tweaked home intro

* Changed advert sections and where they are included

* Fixed forum completed level sections check

* Removed unused files

* Added styles for forum sidebar stats section

* Fixed gramatical error in ticker

* Removed fallback Gravatar image

* Removing unused files

* Removed t-shirt unavailability message from donation page

* Added override for console.log to obfuscate level answers a little

* Added number formating to IRC stats template

* Check for minifier before adding wysiwyg.js

* Moved majority of static resources to CDN

* Removing project status image, service seems to be buggy

* Changed README test

* Added Google Authentication

* Fixed errors with 2FA

* Tweaked wording on enter 2FA code login box

* Added title to entry on Google Auth

* Changed 2FA field to number for easier input on mobile devices

* Added extra forum checks

Don't allow users to post in the first 10 mins of being a member and
don't allow users to post on more than 2 threads, per hour for the first
24 hours

* Fixed mistake in forum verification

* Moved fetch outside if, fixing problem restricting all users posting in the fourm

* Make user.googleAuth function more flexible

* Disable old users from using expired password hashes

* Added online status indicator to levels that need it

* Remove ability to use old password hashes

* Fixed merge

* Fixed warnings on level pages

* Fixed warnings on levels

* Fixed warning on forum pages

* fix(cache): cache wasn't loading the file contents of cache files

* fix(levels): caching wasn't working correctly

* fix(levels): cache wasn't working

* fix(levels): level caching

* fix(settings): fixing warning

* Added fields to SQL to track details on moderator flags

* feat(admin): Improved admin access to user access controls

* fix(admin) timestamps in moderator logs

* Fixing mod_reports table definition to pass installation

* Adding cdns to content security policy so that site will run locally (hackthis-10af.kxcdn.com cdn.socket.io and d3t63m1rxnixd2.cloudfront.net)

* Installation requires ldap support for php or server won't work (no registeration etc.)

* Fix(install): Fixing 500 internal server error (Apache2 on Ubuntu)
  *  Fixes: Invalid command 'RewriteEngine' #167

* Update view.php

line 105, a bug in the code doesn't let a user edit a submitted but not approved article (redirects always to $id=1 and not the correct one)
<?=$article->id;?> instead of <?=$id;?> should do the job?

* Allowing use of 'data:' for pictures

As we can see in console, there is a problem with a script trying to injecting a picture using 'data:' format
The script is this one : https://github.com/HackThis/hackthis.co.uk/blob/master/html/files/js/favcounter.js

The problem is from CSP rule : 'img-src *' which disallow the use of 'data:' format.
With this new rule, the script should works as he has to !

* Added spoiler tag

Spoiler tag has been added with eye icon 
<li class="mobile-hide"><a href="#" class="hint--top" data-hint="Spoiler" data-tag="Spoiler"><i class="icon-eye"></i></a></li>

* Bug fix on user name

itemprop='author' href='/user/flabbyrabbit>'>flabbyrabbit</a>
to 
itemprop='author' href='/user/flabbyrabbit'>flabbyrabbit</a>

There was a extra > symbol which leads to error page

* Bug fix on user name #2

There is a extra > symbol after the username

* Bug fix on user name #3

There is an extra > symbol after the user name

* Update wysiwyg.php

* Update faq.php

* Update faq.php

* Update class.admin.php

* Fixed mismatched close thread CSRF key

* Update version_history.md

Simple spelling error

+++ html/settings/account.php
@@ -7,7 +7,7 @@
 
     if (isset($_GET['password'])) {
         $changed = $app->user->changePassword($_POST['newpassword1'], $_POST['newpassword1']);
-    } else if (isset($_GET['delete'])) {
+    } else if (isset($_GET['delete']) && isset($_POST['delete']) && isset($_POST['token'])) {
         $status = $app->user->delete($_POST['delete'], $_POST['token']);
 
         if ($status === true) {
